+++ src/Symfony/Component/Workflow/Tests/SupportStrategy/InstanceOfSupportStrategyTest.php
@@ -3,27 +3,21 @@
 namespace Symfony\Component\Workflow\Tests\SupportStrategy;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\Workflow\SupportStrategy\ClassInstanceSupportStrategy;
+use Symfony\Component\Workflow\SupportStrategy\InstanceOfSupportStrategy;
 use Symfony\Component\Workflow\Workflow;
 
-/**
- * @group legacy
- */
-class ClassInstanceSupportStrategyTest extends TestCase
+class InstanceOfSupportStrategyTest extends TestCase
 {
-    /**
-     * @expectedDeprecation "Symfony\Component\Workflow\SupportStrategy\ClassInstanceSupportStrategy" is deprecated since Symfony 4.1. Use "Symfony\Component\Workflow\SupportStrategy\InstanceOfSupportStrategy" instead.
-     */
     public function testSupportsIfClassInstance()
     {
-        $strategy = new ClassInstanceSupportStrategy('Symfony\Component\Workflow\Tests\SupportStrategy\Subject1');
+        $strategy = new InstanceOfSupportStrategy(Subject1::class);
 
         $this->assertTrue($strategy->supports($this->createWorkflow(), new Subject1()));
     }
 
     public function testSupportsIfNotClassInstance()
     {
-        $strategy = new ClassInstanceSupportStrategy('Symfony\Component\Workflow\Tests\SupportStrategy\Subject2');
+        $strategy = new InstanceOfSupportStrategy(Subject2::class);
 
         $this->assertFalse($strategy->supports($this->createWorkflow(), new Subject1()));
     }
@@ -34,4 +28,11 @@ class ClassInstanceSupportStrategyTest extends TestCase
             ->disableOriginalConstructor()
             ->getMock();
     }
+}
+
+class Subject1
+{
+}
+class Subject2
+{
 }